Series by power series, graphs drawn, and what the sweeps found
Series used to differentiate its way to the coefficients, and the
derivative of a quotient or a root doubled in size with every order:
Series[Tan[x], {x, 0, 11}], Series[1/(1 + x^2), {x, 0, 10}] and
Series[Sqrt[1 + x], {x, 0, 15}] never came back, and the window froze
with them.  It now works in truncated power series -- arrays of exact
coefficients in x - a, added, multiplied, divided, raised and composed
to the order asked for -- so those come back at once with exact
fractions, Sin[x]/x and x/(Exp[x] - 1) have series at all, Sec[x] does
too, and the expansion point is kept as written, (x - Pi/2)^2/2.  The
derivatives stay for a function the arithmetic does not know, with a
guard on their size rather than a hang.

A graph in the other sense of the word is drawn: Graph[{1 -> 2,
2 -> 3}], AdjacencyGraph[m], MATLAB's digraph(s, t) and graph(s, t),
with TopologicalSort beside them.  A directed acyclic graph is laid
out in layers from its sources down, the vertices in a row ordered
under the average of their neighbours; anything else is laid out by
springs.  The showcase has one, and the README's picture shows it.

Wrong answers the sweeps in CLAUDE.md turned up, each fixed:
InverseLaplaceTransform of (s - 2)/((s - 2)^2 + 9) had a sine it
never had, s/(s + 3)^2 had the wrong sign, and 1/(s + 3)^3 came out as
nonsense with roots at -3.0000027 -- the quadratic case had two signs
wrong and a repeated root above degree two was never divided out
properly.  Total, Plus and Times refused complex numbers.  MatchQ and
Cases compared a list pattern such as {_, _} or {___Integer} as text
and matched nothing.  Mod[10^40, 2^62] came out negative, and
Mod[10^30, 4611686018427387903] lost its last digit on the way to a
double.  1/(-2^62*2) printed as -1/-9223372036854775808.  A non-ASCII
character in the input made an error message the window could not
draw.  math42-calc cut a line past 4095 bytes in two.  A .mat file
claiming more bytes than it held asked for that much memory or read
past its buffer.

MATLAB's if (x > 0), ... end and while (cond), ... end -- the
condition in brackets -- were read as calls of a function named if.
A block laid out over several lines in a .m file was run a line at a
time, so the body ran whatever the condition said; the lines up to
the matching end are joined now.  f @ g @ x chains.  Coth, Sech and
Csch are functions.

In the window: the up and down arrows never walked the history,
because nothing was ever put in it.  Plot[10^16 + Sin[x], {x, 0, 1}]
hung the window in the tick loop, where a step of 0.5 is below the
spacing of the doubles.  Not[x] was set as two wrong bytes.  Opening
a file over a notebook went on counting from where the old one left
off.  --print-to did not mark the process unique.  A file dialog
answering after its window was closed used the freed session.  A
paper size and --convert's path leaked.  A negative number in front
of a product is set as -3 t rather than (-3) t, and a fraction in a
series goes under the power, (x - 1)^2/8, rather than being
multiplied out.

Checked: meson compile with --werror is clean; the CI smoke test
passes; every example gives the same output through math42-calc as
before and round trips through .m, .wl and .nb unchanged; the four
sweeps from CLAUDE.md -- Simplify and D[Integrate] against the
expression's value at 0.37, both spellings of every function, and
fifteen thousand calls of every name with a dozen shapes of argument
-- come back with nothing but the x - x cases and the random
functions; a Series sweep against the function's value at 0.03 is
clean; roots go back into their equations, inverses times their
matrices, Laplace transforms there and back, and integrals against
NIntegrate all come out as zero; the showcase was photographed with
--screenshot for the README and exported to a PDF whose pages carry
the graph.

+++ src/core/m42-bigint.c
@@ -310,22 +310,25 @@ M42Big *
 m42_big_divide_small (const M42Big *a, gint64 d, gint64 *remainder)
 {
   M42Big *out;
-  gint64 carry = 0;
-  gint64 size = d < 0 ? -d : d;
+  guint64 carry = 0;
+  guint64 size = d < 0 ? -(guint64) d : (guint64) d;
 
   if (d == 0)
     return NULL;
   out = big_alloc (a->len);
   for (guint i = a->len; i > 0; i--)
     {
-      gint64 current = carry * M42_BIG_BASE + a->digit[i - 1];
+      /* The carry is below the divisor, which may be anything up to
+       * 2^63, so the step needs more than 64 bits: Mod[10^40, 2^62]
+       * came out negative when it was done in a gint64. */
+      unsigned __int128 current = (unsigned __int128) carry * M42_BIG_BASE + a->digit[i - 1];
 
       out->digit[i - 1] = (guint32) (current / size);
-      carry = current % size;
+      carry = (guint64) (current % size);
     }
   out->sign = a->sign == 0 ? 0 : (d < 0 ? -a->sign : a->sign);
   if (remainder != NULL)
-    *remainder = a->sign < 0 ? -carry : carry;
+    *remainder = a->sign < 0 ? -(gint64) carry : (gint64) carry;
   return big_trim (out);
 }
